Move lib.rs contents to daemon.rs
